% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abs_path}
\alias{abs_path}
\title{This functions creates a absolute file path with forward slashes. It
takes in as its argument a relative path to the data file you want to
read. From that relative path it creates a absolute path to that data
file that is unique to the local file system. This helps avoid file path
errors that people run into when they read each others code.}
\usage{
abs_path(relPathToDataFile)
}
\arguments{
\item{relPathToDataFile}{relative path to the data file (using forward slashes)}
}
\value{
\code{abs_path} absolute path to data file (with forward slashes)
}
\description{
This functions creates a absolute file path with forward slashes. It
takes in as its argument a relative path to the data file you want to
read. From that relative path it creates a absolute path to that data
file that is unique to the local file system. This helps avoid file path
errors that people run into when they read each others code.
}
\examples{
abs_path <- function(relPathToDataFile = "path/to/data.csv")

}
